Allow create folder when creating/writing to a file (#5220)

+++ aws/src/main/java/nl/nn/adapterframework/filesystem/AmazonS3FileSystem.java
@@ -259,6 +259,11 @@ public class AmazonS3FileSystem extends FileSystemBase<S3Object> implements IWri
 
 	@Override
 	public OutputStream createFile(final S3Object f) throws FileSystemException, IOException {
+		String folder = getParentFolder(f);
+		if (folder != null && !folderExists(folder)) { //AWS Supports the creation of folders, this check is purely here so all FileSystems have the same behavior
+			throw new FileSystemException("folder ["+folder+"] does not exist");
+		}
+
 		final File file = FileUtils.createTempFile(".s3-upload");
 		final FileOutputStream fos = new FileOutputStream(file);
 		return new BufferedOutputStream(fos) {
